Moved command end points to LocationsController
