+++ src/EnvironmentMonitor.WebApi/Controllers/DeviceCommandsController.cs
@@ -14,7 +14,6 @@ namespace EnvironmentMonitor.WebApi.Controllers
     public class DeviceCommandsController : ControllerBase
     {
         private readonly IDeviceCommandService _deviceCommandService;
-        private readonly ILocationCommandService _locationCommandService;
         private readonly ILogger<DeviceCommandsController> _logger;
 
         public DeviceCommandsController(
@@ -30,14 +29,5 @@ namespace EnvironmentMonitor.WebApi.Controllers
         [HttpGet("{deviceIdentifier}/attributes")]
         public async Task<Dictionary<int, string>> GetDeviceAttributes([FromRoute] string deviceIdentifier) => await _deviceCommandService.GetDeviceAttributes(deviceIdentifier);
 
-        [HttpPost("location/{locationIdentifier}/motion-control-status")]
-        [Authorize(Roles = "Admin")]
-        public async Task SetLocationMotionControlStatus([FromRoute] Guid locationIdentifier, [FromBody] SetMotionControlStatusMessage model)
-            => await _locationCommandService.SetMotionControlStatus(locationIdentifier, (MotionControlStatus)model.Mode, model.ExecuteAt);
-
-        [HttpPost("location/{locationIdentifier}/motion-control-delay")]
-        [Authorize(Roles = "Admin")]
-        public async Task SetLocationMotionControlDelay([FromRoute] Guid locationIdentifier, [FromBody] SetMotionControlDelayMessag model)
-            => await _locationCommandService.SetMotionControlDelay(locationIdentifier, model.DelayMs, model.ExecuteAt);
     }
 }
